Moved logger and loggerFactory to another directory
